requirements.txt
docker==7.0.0
python-dotenv==1.0.1
py-cord[speed] @ git+https://github.com/Pycord-Development/pycord@8a7ea47
jishaku==2.5.2
# GitPython==3.1.29
humanize==4.9.0
psutil==5.9.8
tortoise-orm[asyncpg]==0.20.0
setuptools==69.1.0
